Add service to the definitions menu
