+++ composer/packages/ast-model/package.json
@@ -19,7 +19,8 @@
     "coverage": "composer pkg:test:coverage",
     "watch": "composer pkg:watch",
     "watch:test": "composer pkg:test:watch",
-    "gen-ast-utils": "npm run build && node lib/tools/generate"
+    "gen-ast-utils": "npm run build && node lib/tools/generate",
+    "gen-default-nodes": "npm run build && node lib/tools/gen-default-nodes"
   },
   "dependencies": {
     "@ballerina/lang-service": "^0.983.1-SNAPSHOT",
